Created folder for user-service,added Dockerfile for that service,added docker-compose
