+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "go.toolsEnvVars": {
+        "GOROOT": "C:\\Users\\Kleopatra\\Desktop\\IV\\XML-BSEP\\XML"
+    }
+}
