+++ XML/docker-compose.yml
@@ -1,0 +1,44 @@
+version: '3.7'
+services:
+  userservice:
+    build: 
+      context: ./user-service/
+      dockerfile: Dockerfile
+    image: userservice1
+    container_name: user_server
+    restart: always
+    environment:
+      - HOST=database
+      - PG_DBPORT=5432
+      - PG_USER=postgres
+      - PG_PASSWORD=postgres
+      - XML_DB_NAME=postgres
+    networks:
+      - servers
+    ports:
+      - 8082:8082
+    depends_on: 
+      - database
+
+  database:  
+    image: postgres:14.1-alpine
+    restart: always
+    networks:
+      - servers
+    environment:
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+    ports:
+      - '5432:5432'
+    volumes: 
+      - database-data:/var/lib/postgresql/data
+
+
+volumes:
+  database-data:
+    name: server-database
+
+networks:
+  servers:
+    name: servers
+    driver: bridge
