--- XML/dto/LogInResponse.go
@@ -1,5 +1,0 @@
-package dto
-
-type LogInResponseDto struct {
-	Token string
-}
