--- XML/dto/LoginRequest.go
@@ -1,6 +1,0 @@
-package dto
-
-type LoginRequest struct {
-	Username string `json:"username" form:"username" binding:"required"`
-	Password string `json:"password" form:"password" binding:"required"`
-}
