--- XML/model/Gender.go
@@ -1,9 +1,0 @@
-package model
-
-type Gender int
-
-const (
-	MALE Gender = iota
-	FEMALE
-	OTHER
-)
