--- XML/model/UserType.go
@@ -1,9 +1,0 @@
-package model
-
-type UserType int
-
-const (
-	ADMIN UserType = iota
-	REGISTERED_USER
-	AGENT
-)
