+++ XML/user-service/dto/LogInResponse.go
@@ -1,0 +1,5 @@
+package dto
+
+type LogInResponseDto struct {
+	Token string
+}
