+++ XML/user-service/dto/LoginRequest.go
@@ -1,0 +1,6 @@
+package dto
+
+type LoginRequest struct {
+	Username string `json:"username" form:"username" binding:"required"`
+	Password string `json:"password" form:"password" binding:"required"`
+}
