+++ XML/user-service/go.mod
@@ -1,0 +1,42 @@
+module user/module
+
+go 1.18
+
+require github.com/google/uuid v1.3.0
+
+require github.com/go-sql-driver/mysql v1.6.0
+
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/lib/pq v1.10.5
+)
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/labstack/echo v3.3.10+incompatible
+)
+
+require (
+	github.com/euroteltr/rbac v1.0.9
+	github.com/jackc/chunkreader/v2 v2.0.1 // indirect
+	github.com/jackc/pgconn v1.12.0 // indirect
+	github.com/jackc/pgio v1.0.0 // indirect
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgproto3/v2 v2.3.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20200714003250-2b9c44734f2b // indirect
+	github.com/jackc/pgtype v1.11.0 // indirect
+	github.com/jackc/pgx/v4 v4.16.0 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.4 // indirect
+	github.com/labstack/gommon v0.3.1 // indirect
+	github.com/mattn/go-colorable v0.1.11 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.1 // indirect
+	golang.org/x/crypto v0.0.0-20220427172511-eb4f295cb31f
+	golang.org/x/net v0.0.0-20211112202133-69e39bad7dc2 // indirect
+	golang.org/x/sys v0.0.0-20211103235746-7861aae1554b // indirect
+	golang.org/x/text v0.3.7 // indirect
+	gorm.io/driver/postgres v1.3.5
+	gorm.io/gorm v1.23.5
+)
