+++ XML/user-service/model/Gender.go
@@ -1,0 +1,9 @@
+package model
+
+type Gender int
+
+const (
+	MALE Gender = iota
+	FEMALE
+	OTHER
+)
