+++ XML/user-service/model/UserType.go
@@ -1,0 +1,9 @@
+package model
+
+type UserType int
+
+const (
+	ADMIN UserType = iota
+	REGISTERED_USER
+	AGENT
+)
